Create autostart default network.
This is (always was) the default in Upstream as well.
But Debian disables the upstream habit of enabling it by default.

Since recent upstream the following former Ubuntu custom changes are default:
- nat only on some ports <port start='1024' end='65535'/>

Ubuntu still behaves different in not always enabling the default network.
Cases we do not enable it are when 192.168.122.1/24 is already taken.

We evaluated to drop this, but we want to keep the default enabling of
the default network to provide users the smothest start into using
libvirt.
Instead of dropping a few refreshes and minor extensions to the
functionality took place in 18.04 adding:
 - bug fixups on sed call
 - use modern shell syntax
 - make ip to check an argument
 - try more potential default networks
 - since we add it, also remove it on purge

The former code contained sections that only were relevant when
upgrading from old versions that had a the symlink owned by the package.
After 18.04 these sections were dropped.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>