UBUNTU-only: fixup smoke-lxc by ignoring potential issues on destroy as long as undefine succeeds
virsh destroy on a libvirt-lxc domain might return bad RC but still
work. The following undefine would not work if it would be still
running.
Since on Ubuntu lbivirt-lxc is discouraged anyway in favor of lxd let us
ignore the bad RC of the destroy call to get the tests reliable.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>