systemd: avoid --no-start which breaks .sockets on re-install
dh_installsystemd's --no-start is conflicting with .socket files
as after their removal without being stopped they will enter a faulty
state that can only be recovered with a restart of the socket.

That restart never happens which renders the socket unusable on actions
like install/remove/install.

--no-start was added there for other reasons and no more needed with
just the sockets. This also names the group accordingly to reflect that
these are now just the sockets.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>