systemd: --no-restart-after-upgrade does not prevent restarts
While it sounds that way `--no-restart-after-upgrade` does not prevent
restarts. It only controls if the restart happens after the full upgrade
(default in new compat levels) OR with this option being set if it is
stopped in prerm and started in postinst (longer downtime but using
stop/start instead of restart which can sometimes be beneficial).

The proper --no-stop-on-upgrade was dropped by accident in ff981d56f and
later replaced with the wrong --no-restart-after-upgrade. But all that
was done for libvirt-guests.service and isn't needed for the sockets.

Fixes: ff981d56f "Pass --no-auto to dh_instalsystemd"
Fixes: a5dd08c65 "d/rules: systemd: Also pass --no-restart-on-upgrade when
                  using --no-start"

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>