systemd: let sockets use --no-stop-on-upgrade to avoid false positives
When upgrading (or re-installing) the default dh_installsystemd snippet
will now contain a section that will try to restart the socket(s).
But if the corresponding services are already running that will trigger
  Job failed. See "journalctl -xe" for details.
The reason for that is said already running service:
  libvirtd-admin.socket: Socket service libvirtd.service already active, refusing.
  Failed to listen on Libvirt admin socket.

Do not touch the .socket units on upgrade to avoid this.

The only drawback is that if we ever deploy a change/fix to sockets it will
not be automatically reloaded/restarted. But this issue was already present
before this series.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>