control: Always use versioned deps for libvirt components
libvirt packaging is fairly modularized because we want users
to be able to tailor the installation to their needs by deciding
which components to include and which ones to leave out, but
mixing and matching different versions is very much not intended
to work.

Right now we're not always enforcing this tight coupling, and
that can result in failures at runtime, especially during
upgrades. Let's address the problem by making sure all
dependencies between libvirt components are versioned.

Closes: #981435